requirements.txt
# anyio==3.5.0
# asgiref==3.5.0
# click==8.0.4
# fastapi==0.68.2
# h11==0.13.0
# httptools==0.2.0
# idna==3.3
# joblib==1.1.0
# nltk==3.7
# numpy==1.22.3
# pydantic==1.9.0
# python-dotenv==0.19.2
# PyYAML==6.0
# regex==2022.3.15
# sniffio==1.2.0
# starlette==0.14.2
# tqdm==4.63.0
# typing-extensions==4.1.1
# uvicorn==0.15.0
# uvloop==0.16.0
# watchgod==0.8.1
# websockets==10.2

anyio==3.5.0
arxiv==1.4.2
asgiref==3.5.0
avro==1.11.0
boto3==1.21.28
botocore==1.24.28
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
configparser==5.2.0
cryptography==36.0.2
decorator==5.1.1
docopt==0.6.2
fastapi==0.68.2
fastavro==1.4.10
feedparser==6.0.8
gssapi==1.7.3
h11==0.13.0
hdfs==2.6.0
httptools==0.2.0
idna==3.3
jmespath==1.0.0
joblib==1.1.0
kerberos==1.3.1
krb5==0.3.0
nltk==3.7
ntlm-auth==1.5.0
numpy==1.22.3
pandas==1.4.1
pdfminer.six==20220319
pycparser==2.21
pydantic==1.9.0
pykerberos==1.2.4
pyspnego==0.5.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
pywinrm==0.4.2
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-kerberos==0.14.0
requests-ntlm==1.1.0
s3transfer==0.5.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.8.1
websockets==10.2
xmltodict==0.12.0